tcp: Move dynamic metrics handling into separate file.
